+++ services/game_services.py
@@ -3,7 +3,9 @@ from chess_app.chess_engine import *
 from custom_errors.custom_errors import *
 from database_services.strapi_api_service import StrapiApiService
 from typing import Optional, List
+from services.chess_engine_services import *
 from services.player_services import *
+
 
 api_service = StrapiApiService()
 
@@ -110,29 +112,3 @@ def search_and_join_game(player_name: str):
         )
     updated_game = api_service.update_game_with_new_player(player, available_game[0])
     return updated_game
-
-
-def init_game(game_uuid: str):
-    """Starts a Game if two players joined the Game, init an FEN code to the Game instance
-
-    Args:
-        game_uuid (str): Uuid of the Game
-    """
-    game = retrieve_single_game(game_uuid)
-    if game.fen:
-        raise GameAlreadyStartedError(f"Game already started with fen : {game.fen}")
-    if len(game.players) != 2:
-        raise NotActiveGameError(
-            f"Not enough Players in the Game: {len(game.players)} joined"
-        )
-    try:
-        player_1 = game.players[0].name
-        player_2 = game.players[1].name
-        chess_engine = GameOfChess(name_player_1=player_1, name_player_2=player_2)
-        fen = chess_engine.return_fen()
-        game.is_active = True
-        updated_game = api_service.update_fen_of_game(game=game, fen=fen)
-        return updated_game
-
-    except GameNotFoundError:
-        raise GameNotFoundError(f"No game found with UUID: {game_uuid}")
